refactor: wrap into getItemIdentifier, internalize other functions
