+++ packages/common/src/items/index.ts
@@ -22,6 +22,7 @@ export * from "./create-item-from-url";
 export * from "./create-item-from-url-or-file";
 export * from "./uploadImageResource";
 export * from "./is-services-directory-disabled";
+export * from "./getItemIdentifier";
 // No longer exported as the only App Hub needed this for was a Site
 // and that registration is now done in the domain service with a
 // signed HMAC request.
