+++ packages/common/src/items/slugs.ts
@@ -2,34 +2,7 @@ import { getItem, ISearchOptions, searchItems } from "@esri/arcgis-rest-portal";
 import { IRequestOptions } from "@esri/arcgis-rest-request";
 import { IItem } from "@esri/arcgis-rest-types";
 import { slugify } from "../utils";
-
-// TODO: work out how to unify content slug fns
-// https: github.com/Esri/hub.js/blob/master/packages/common/src/content/index.ts#L301-L348
-
-/**
- * Uri slugs have `::` as a separator, but we need to use `|` in the typeKeywords. This function
- * converts a uri slug to a typeKeyword slug.
- * @param slug
- * @returns
- */
-export function uriSlugToKeywordSlug(slug: string): string {
-  if (slug.indexOf("::") > -1) {
-    slug = slug.replace("::", "|");
-  }
-  return slug;
-}
-
-/**
- * Convert a typeKeyword slug to a uri slug. This is the reverse of uriSlugToKeywordSlug
- * @param slug
- * @returns
- */
-export function keywordSlugToUriSlug(slug: string): string {
-  if (slug.indexOf("|") > -1) {
-    slug = slug.replace("|", "::");
-  }
-  return slug;
-}
+import { uriSlugToKeywordSlug } from "./_internal/slugConverters";
 
 /**
  * Create a slug, namespaced to an org
